show core snackbar when user publish post for the first time as part of onboarding
