+++ apps/wpcom-block-editor/src/wpcom/features/redirect-onboarding-user-after-publishing-post.js
@@ -71,52 +71,3 @@ export function RedirectOnboardingUserAfterPublishingPost() {
 		}
 	} );
 }
-
-export function RedirectToHomeLaunchpadAfterPublishingPost() {
-	const currentPostType = useSelect(
-		( localSelect ) => localSelect( 'core/editor' ).getCurrentPostType(),
-		[]
-	);
-
-	const hasPublishFirstPostTaskQueryArg =
-		getQueryArg( window.location.search, 'task' ) === 'publish_first_post';
-
-	if ( ! hasPublishFirstPostTaskQueryArg || currentPostType !== 'post' ) {
-		return false;
-	}
-
-	// Save site origin in session storage to be used in editor refresh.
-	const siteOriginParam = getQueryArg( window.location.search, 'origin' );
-	if ( siteOriginParam ) {
-		window.sessionStorage.setItem( 'site-origin', siteOriginParam );
-	}
-
-	const siteOrigin = window.sessionStorage.getItem( 'site-origin' ) || 'https://wordpress.com';
-	const siteSlug = window.location.hostname;
-
-	const unsubscribe = subscribe( () => {
-		const isSavingPost = select( 'core/editor' ).isSavingPost();
-		const isCurrentPostPublished = select( 'core/editor' ).isCurrentPostPublished();
-		const isCurrentPostScheduled = select( 'core/editor' ).isCurrentPostScheduled();
-		//const getCurrentPostRevisionsCount = select( 'core/editor' ).getCurrentPostRevisionsCount();
-		//const notices = select( 'core/notices' ).getNotices();
-		if ( ! isSavingPost && ( isCurrentPostPublished || isCurrentPostScheduled ) ) {
-			unsubscribe();
-
-			dispatch( 'core/edit-post' ).closePublishSidebar();
-
-			// Show success notice with Next steps link
-			dispatch( 'core/notices' ).createSuccessNotice( 'Well done publishing your first post!', {
-				actions: [
-					{
-						label: 'Next steps',
-						url: `${ siteOrigin }/home/${ siteSlug }`,
-					},
-				],
-				type: 'snackbar',
-				isDismissible: true,
-				id: 'NEXT_STEPS_NOTICE_ID',
-			} );
-		}
-	} );
-}
